pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manim_code_blocks_redux"
version = "1.0.13"
description = "Animate highlighted blocks of code in Manim"
readme = "README.md"
authors = [{ name = "chand1012" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["tokenize-all-code"]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/CasualCodersProjects/manim-code-blocks"
